deleted ProductType and changed Product and Recipe repositories
