+++ CashaWeb/Casha.Core/DbModels/Product.cs
@@ -12,10 +12,6 @@ namespace Casha.Core.DbModels
 
         public string Name { get; set; }
 
-        public int ProductTypeId { get; set; }
-
-        public ProductType ProductType { get; set; }
-
         public List<RecipeProduct> RecipeProducts { get; set; }
         
         public List<UserProduct> UserProducts { get; set; }
